+++ lib/board.rb
@@ -1,0 +1,18 @@
+class Board
+  def initialize
+    @the_board = {
+      1 => ' ', 2 => ' ', 3 => ' ',
+      4 => ' ', 5 => ' ', 6 => ' ',
+      7 => ' ', 8 => ' ', 9 => ' '
+    }
+  end
+
+  def restart_game
+    @the_board = {
+      1 => ' ', 2 => ' ', 3 => ' ',
+      4 => ' ', 5 => ' ', 6 => ' ',
+      7 => ' ', 8 => ' ', 9 => ' '
+    }
+  end
+
+end
